data helper refactor; added logger helper
