+++ Helper/Logger.php
@@ -1,0 +1,75 @@
+<?php
+namespace Klaviyo\Reclaim\Helper;
+
+use Magento\Framework\Filesystem\DirectoryList;
+use Klaviyo\Reclaim\Logger\Logger as KlaviyoLogger;
+use Klaviyo\Reclaim\Helper\ScopeSetting;
+
+class Logger extends \Magento\Framework\App\Helper\AbstractHelper
+{
+    /**
+     * directory list interface
+     * used to programmatically retrieve paths within magento app install
+     * @var DirectoryList
+     */
+    protected $_dir;
+
+    /**
+     * path to the log file
+     * @var string
+     */
+    protected $_logPath;
+
+    /**
+     * klaviyo logger object
+     * @var KlaviyoLogger
+     */
+    protected $_klaviyoLogger;
+
+    /**
+     * is the logger enabled?
+     * @var boolean
+     */
+    protected $_loggerEnabled;
+
+    public function __construct(
+        DirectoryList $dir,
+        KlaviyoLogger $klaviyoLogger,
+        ScopeSetting $klaviyoScopeSetting
+    )
+    {
+        $this->_dir = $dir;
+        $this->_klaviyoLogger = $klaviyoLogger;
+        $this->_loggerEnabled = $klaviyoScopeSetting->getLoggerEnabled();
+        $this->_logPath = $this->_dir->getPath('log') . '/klaviyo.log';
+    }
+
+    /**
+     * Getter method for the logfile's path
+     * @return string
+     */
+    public function getPath()
+    {
+        return $this->_logPath;
+    }
+
+    /**
+     * Method to log the provided message
+     * @param string $message
+     */
+    public function log($message)
+    {
+        if ($this->loggerEnabled) {
+            $this->_klaviyoLogger->info($message);
+        }
+    }
+
+    /**
+     * do I need this? who tf knows
+     * probably
+     */
+    public function clean($since)
+    {
+
+    }
+}
